move losses to utils
